bullet disappearance, damage ready

+++ background.cpp
@@ -8,7 +8,7 @@ Background::Background(sf::Vector2f size, sf::Vector2f texture_size, std::string
     setTexture(&texture_);
 }
 
-void Background::repeate(int x, int y)
+void Background::repeate(float x, float y)
 {
     setTextureRect( sf::IntRect(0, 0, texture_size.x*x, texture_size.y*y));
 }
